trading_backend/loc-service/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
pydantic
httpx
websockets
python-jose[cryptography]